feat(plan): add migration preview endpoint
## What

- Added a non-mutating `PlanMigrations` RPC that reports current status, latest source version, target version, direction, and pending up migration versions for configured databases.
- Wired the new contract through generated Go/Ruby protobuf artifacts, core migration inspection, versioned API mapping, HTTP RPC routing, and gRPC transport handling.
- Added HTTP and gRPC feature coverage for pending migrations and already-current databases, plus README and API comments documenting the informational boundary and non-goals.

## Why

- Operators can now inspect what Migrieren would advance toward before applying migrations, without exposing source or database URLs and without introducing step-based, rollback, all-down, or repair semantics.
- The API remains additive and compatibility-safe while preserving the existing `Migrate` and `ApplyMigrations` execution boundaries.

## Testing

- `make proto-generate` - passed
- `make proto-lint` - passed
- `make proto-breaking` - passed
- `go test ./...` - passed
- `make lint` - passed
- `make features` - passed; covers the new HTTP and gRPC planning scenarios along with the existing feature suite.

AI-assisted-by: [Codex](https://openai.com/codex/)
