+++ api/migrieren/v1/service.proto
@@ -87,6 +87,59 @@ message MigrationStatus {
   MigrationState state = 3;
 }
 
+// MigrationDirection identifies whether the current database state can move
+// toward the latest available source version by applying up migrations.
+enum MigrationDirection {
+  // MIGRATION_DIRECTION_UNSPECIFIED is the default value for unknown directions.
+  MIGRATION_DIRECTION_UNSPECIFIED = 0;
+
+  // MIGRATION_DIRECTION_NONE means no up migrations are pending.
+  MIGRATION_DIRECTION_NONE = 1;
+
+  // MIGRATION_DIRECTION_UP means one or more up migrations are pending.
+  MIGRATION_DIRECTION_UP = 2;
+}
+
+// MigrationPlan reports non-mutating migration planning information for a
+// configured database.
+message MigrationPlan {
+  // status reports the current database migration version state.
+  MigrationStatus status = 1;
+
+  // latest_version is the highest migration version available from the
+  // configured source. It is zero when the source has no migrations.
+  uint64 latest_version = 2;
+
+  // target_version is the version a clean or unapplied database can converge
+  // toward using pending up migrations. It equals latest_version when direction
+  // is MIGRATION_DIRECTION_UP, and otherwise equals the current status version.
+  uint64 target_version = 3;
+
+  // direction reports whether pending up migrations can currently apply.
+  MigrationDirection direction = 4;
+
+  // pending_versions contains source versions greater than the current status
+  // version, in apply order. It is empty when no up migrations are pending.
+  repeated uint64 pending_versions = 5;
+}
+
+// PlanMigrationsRequest identifies the configured database to inspect against
+// its migration source.
+message PlanMigrationsRequest {
+  // database is the configured logical database name, not a raw database URL.
+  string database = 1;
+}
+
+// PlanMigrationsResponse contains response metadata and the migration plan.
+message PlanMigrationsResponse {
+  // meta contains request metadata emitted by the service runtime.
+  map<string, string> meta = 1;
+
+  // plan reports current status and pending source versions without applying
+  // migrations.
+  MigrationPlan plan = 2;
+}
+
 // StatusRequest identifies the configured database to inspect.
 message StatusRequest {
   // database is the configured logical database name, not a raw database URL.
@@ -158,6 +211,22 @@ service Service {
   // migrate v4 context support, which is currently incomplete in some database
   // driver paths.
   rpc ApplyMigrations(ApplyMigrationsRequest) returns (ApplyMigrationsResponse) {}
+
+  // PlanMigrations reports current status and pending up migration versions for
+  // a configured database without applying migration files.
+  //
+  // Errors use NotFound for unknown databases, Canceled/DeadlineExceeded for
+  // stopped requests, and Internal for configuration, source, or database
+  // inspection failures.
+  //
+  // Failures after request routing expose the same safe diagnostic trailers as
+  // Migrate, including migration-error, migration-log-count, migration-stage,
+  // and migration-log-last when those values are available.
+  //
+  // PlanMigrations opens the configured migration source to discover available
+  // versions, but it does not read migration bodies or expose configured source
+  // strings, database URL strings, or resolved secret values.
+  rpc PlanMigrations(PlanMigrationsRequest) returns (PlanMigrationsResponse) {}
 
   // Status reports the current migration version state for a configured
   // database.
