--- test/features/step_definitions/v1/transport/http/api_steps.rb
@@ -1,135 +1,0 @@
-# frozen_string_literal: true
-
-When('I request to migrate with HTTP:') do |table|
-  @response = request_with_http(table)
-end
-
-When('I request to apply migrations with HTTP:') do |table|
-  @response = request_apply_with_http(table)
-end
-
-When('I request migration status with HTTP:') do |table|
-  @response = request_status_with_http(table)
-end
-
-When('I request configured databases with HTTP') do
-  @response = request_databases_with_http
-end
-
-Then('I should receive a successful migration from HTTP:') do |table|
-  expect(@response.code).to eq(200)
-
-  resp = JSON.parse(@response.body)
-  migration = resp['migration']
-  rows = table.rows_hash
-  logs = migration['logs'] || []
-
-  expect(resp['meta'].length).to be > 0
-  expect(migration['database']).to eq(rows['database'])
-  expect(migration['version']).to eq(rows['version'].to_i)
-  expect(logs.length).to be >= 0
-
-  expect_postgres_migration(rows['version'].to_i) if rows['database'] == 'postgres'
-  expect_log_migration(rows['version'].to_i) if rows['database'] == 'logs'
-end
-
-Then('I should receive a not found migration from HTTP') do
-  expect(@response.code).to eq(404)
-end
-
-Then('I should receive a migration status from HTTP:') do |table|
-  expect(@response.code).to eq(200)
-
-  resp = JSON.parse(@response.body)
-  status = resp['status']
-  rows = table.rows_hash
-
-  expect(resp['meta'].length).to be > 0
-  expect(status['database']).to eq(rows['database'])
-  expect(status.fetch('version', 0)).to eq(rows['version'].to_i)
-  expect(migration_state(status['state'])).to eq(rows['state'])
-end
-
-Then('I should receive configured databases from HTTP:') do |table|
-  expect(@response.code).to eq(200)
-
-  resp = JSON.parse(@response.body)
-
-  expect(resp['meta'].length).to be > 0
-  expect(resp['databases'].map { |database| database['name'] }).to eq(table.hashes.map { |row| row['database'] })
-end
-
-Then('I should receive an invalid argument migration from HTTP') do
-  expect(@response.code).to eq(400)
-end
-
-Then('I should receive an invalid migration from HTTP') do
-  expect(@response.code).to eq(500)
-end
-
-Then('I should receive failure diagnostics from HTTP:') do |table|
-  rows = table.rows_hash
-  log_count = http_header_value(@response, 'migration-log-count').to_i
-
-  expect(http_header_value(@response, 'migration-error')).to eq(rows['error'])
-  expect(http_header_value(@response, 'migration-stage')).to eq(empty_to_nil(rows['stage']))
-
-  if rows['logs'] == 'present'
-    expect(log_count).to be > 0
-    expect(http_header_value(@response, 'migration-log-last')).not_to be_empty
-  else
-    expect(log_count).to eq(0)
-    expect(http_header_value(@response, 'migration-log-last')).to be_nil
-  end
-end
-
-def request_with_http(table)
-  rows = table.rows_hash
-  opts = Migrieren.http_options(
-    headers: {
-      user_agent: 'Migrieren-ruby-client/1.0 HTTP/1.0',
-      content_type: :json, accept: :json
-    }
-  )
-
-  Migrieren::V1.server_http.migrate(rows['database'], rows['version'].to_i, opts)
-end
-
-def request_apply_with_http(table)
-  rows = table.rows_hash
-  opts = Migrieren.http_options(
-    headers: {
-      user_agent: 'Migrieren-ruby-client/1.0 HTTP/1.0',
-      content_type: :json, accept: :json
-    }
-  )
-
-  Migrieren::V1.server_http.apply_migrations(rows['database'], opts)
-end
-
-def request_status_with_http(table)
-  rows = table.rows_hash
-  opts = Migrieren.http_options(
-    headers: {
-      user_agent: 'Migrieren-ruby-client/1.0 HTTP/1.0',
-      content_type: :json, accept: :json
-    }
-  )
-
-  Migrieren::V1.server_http.status(rows['database'], opts)
-end
-
-def request_databases_with_http
-  opts = Migrieren.http_options(
-    headers: {
-      user_agent: 'Migrieren-ruby-client/1.0 HTTP/1.0',
-      content_type: :json, accept: :json
-    }
-  )
-
-  Migrieren::V1.server_http.list_databases(opts)
-end
-
-def http_header_value(response, key)
-  response.headers[key.tr('-', '_').to_sym]
-end
