+++ test/features/step_definitions/v1/transport/migration_steps.rb
@@ -34,6 +34,27 @@ def migration_state(state)
   MIGRATION_STATES.fetch(state) { MIGRATION_STATES.fetch(state.to_s) }
 end
 
+MIGRATION_DIRECTIONS = {
+  1 => 'none',
+  2 => 'up',
+  'MIGRATION_DIRECTION_NONE' => 'none',
+  'MIGRATION_DIRECTION_UP' => 'up'
+}.freeze
+
+def migration_direction(direction)
+  MIGRATION_DIRECTIONS.fetch(direction) { MIGRATION_DIRECTIONS.fetch(direction.to_s) }
+end
+
+def migration_versions(versions)
+  return Range.new(*versions.split('..').map(&:to_i)).to_a if versions.include?('..')
+
+  versions.split(',').reject(&:empty?).map(&:to_i)
+end
+
+def empty_to_nil(value)
+  value.empty? ? nil : value
+end
+
 Then('I should not see a completed timeout migration') do
   expect(Migrieren.pg.timeout_migration_version).to be_nil
 end
