+++ test/lib/migrieren/v1/http.rb
@@ -48,6 +48,22 @@ module Migrieren
       end
 
       ##
+      # Calls the `PlanMigrations` RPC via the HTTP façade.
+      #
+      # This maps to the HTTP RPC route:
+      # `POST /migrieren.v1.Service/PlanMigrations`
+      #
+      # The request body is JSON with the following shape:
+      # `{ "database": String }`
+      #
+      # @param database [String] logical database name as configured in the service
+      # @param opts [Hash] optional request options passed through to `post`
+      # @return [Object] whatever `Nonnative::HTTPClient#post` returns (typically a response wrapper)
+      def plan_migrations(database, opts = {})
+        post('/migrieren.v1.Service/PlanMigrations', { database: }.to_json, opts)
+      end
+
+      ##
       # Calls the `Status` RPC via the HTTP façade.
       #
       # This maps to the HTTP RPC route:
