+++ test/lib/migrieren/v1/service_pb.rb
@@ -5,7 +5,7 @@
 require 'google/protobuf'
 
 
-descriptor_data = "\n\x1amigrieren/v1/service.proto\x12\x0cmigrieren.v1\"U\n\tMigration\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\x12\x12\n\x04logs\x18\x03 \x03(\tR\x04logs\"F\n\x0eMigrateRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\"\xbe\x01\n\x0fMigrateResponse\x12;\n\x04meta\x18\x01 \x03(\x0b\x32\'.migrieren.v1.MigrateResponse.MetaEntryR\x04meta\x12\x35\n\tmigration\x18\x02 \x01(\x0b\x32\x17.migrieren.v1.MigrationR\tmigration\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"4\n\x16\x41pplyMigrationsRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\"\xce\x01\n\x17\x41pplyMigrationsResponse\x12\x43\n\x04meta\x18\x01 \x03(\x0b\x32/.migrieren.v1.ApplyMigrationsResponse.MetaEntryR\x04meta\x12\x35\n\tmigration\x18\x02 \x01(\x0b\x32\x17.migrieren.v1.MigrationR\tmigration\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"{\n\x0fMigrationStatus\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\x12\x32\n\x05state\x18\x03 \x01(\x0e\x32\x1c.migrieren.v1.MigrationStateR\x05state\"+\n\rStatusRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\"\xbc\x01\n\x0eStatusResponse\x12:\n\x04meta\x18\x01 \x03(\x0b\x32&.migrieren.v1.StatusResponse.MetaEntryR\x04meta\x12\x35\n\x06status\x18\x02 \x01(\x0b\x32\x1d.migrieren.v1.MigrationStatusR\x06status\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"\x1e\n\x08\x44\x61tabase\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\"\x16\n\x14ListDatabasesRequest\"\xc9\x01\n\x15ListDatabasesResponse\x12\x41\n\x04meta\x18\x01 \x03(\x0b\x32-.migrieren.v1.ListDatabasesResponse.MetaEntryR\x04meta\x12\x34\n\tdatabases\x18\x02 \x03(\x0b\x32\x16.migrieren.v1.DatabaseR\tdatabases\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01*\x86\x01\n\x0eMigrationState\x12\x1f\n\x1bMIGRATION_STATE_UNSPECIFIED\x10\x00\x12\x1d\n\x19MIGRATION_STATE_UNAPPLIED\x10\x01\x12\x19\n\x15MIGRATION_STATE_CLEAN\x10\x02\x12\x19\n\x15MIGRATION_STATE_DIRTY\x10\x03\x32\xd8\x02\n\x07Service\x12H\n\x07Migrate\x12\x1c.migrieren.v1.MigrateRequest\x1a\x1d.migrieren.v1.MigrateResponse\"\x00\x12`\n\x0f\x41pplyMigrations\x12$.migrieren.v1.ApplyMigrationsRequest\x1a%.migrieren.v1.ApplyMigrationsResponse\"\x00\x12\x45\n\x06Status\x12\x1b.migrieren.v1.StatusRequest\x1a\x1c.migrieren.v1.StatusResponse\"\x00\x12Z\n\rListDatabases\x12\".migrieren.v1.ListDatabasesRequest\x1a#.migrieren.v1.ListDatabasesResponse\"\x00\x42\x45Z3github.com/alexfalkowski/migrieren/api/migrieren/v1\xea\x02\rMigrieren::V1b\x06proto3"
+descriptor_data = "\n\x1amigrieren/v1/service.proto\x12\x0cmigrieren.v1\"U\n\tMigration\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\x12\x12\n\x04logs\x18\x03 \x03(\tR\x04logs\"F\n\x0eMigrateRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\"\xbe\x01\n\x0fMigrateResponse\x12;\n\x04meta\x18\x01 \x03(\x0b\x32\'.migrieren.v1.MigrateResponse.MetaEntryR\x04meta\x12\x35\n\tmigration\x18\x02 \x01(\x0b\x32\x17.migrieren.v1.MigrationR\tmigration\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"4\n\x16\x41pplyMigrationsRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\"\xce\x01\n\x17\x41pplyMigrationsResponse\x12\x43\n\x04meta\x18\x01 \x03(\x0b\x32/.migrieren.v1.ApplyMigrationsResponse.MetaEntryR\x04meta\x12\x35\n\tmigration\x18\x02 \x01(\x0b\x32\x17.migrieren.v1.MigrationR\tmigration\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"{\n\x0fMigrationStatus\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\x12\x18\n\x07version\x18\x02 \x01(\x04R\x07version\x12\x32\n\x05state\x18\x03 \x01(\x0e\x32\x1c.migrieren.v1.MigrationStateR\x05state\"\xff\x01\n\rMigrationPlan\x12\x35\n\x06status\x18\x01 \x01(\x0b\x32\x1d.migrieren.v1.MigrationStatusR\x06status\x12%\n\x0elatest_version\x18\x02 \x01(\x04R\rlatestVersion\x12%\n\x0etarget_version\x18\x03 \x01(\x04R\rtargetVersion\x12>\n\tdirection\x18\x04 \x01(\x0e\x32 .migrieren.v1.MigrationDirectionR\tdirection\x12)\n\x10pending_versions\x18\x05 \x03(\x04R\x0fpendingVersions\"3\n\x15PlanMigrationsRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\"\xc6\x01\n\x16PlanMigrationsResponse\x12\x42\n\x04meta\x18\x01 \x03(\x0b\x32..migrieren.v1.PlanMigrationsResponse.MetaEntryR\x04meta\x12/\n\x04plan\x18\x02 \x01(\x0b\x32\x1b.migrieren.v1.MigrationPlanR\x04plan\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"+\n\rStatusRequest\x12\x1a\n\x08\x64\x61tabase\x18\x01 \x01(\tR\x08\x64\x61tabase\"\xbc\x01\n\x0eStatusResponse\x12:\n\x04meta\x18\x01 \x03(\x0b\x32&.migrieren.v1.StatusResponse.MetaEntryR\x04meta\x12\x35\n\x06status\x18\x02 \x01(\x0b\x32\x1d.migrieren.v1.MigrationStatusR\x06status\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01\"\x1e\n\x08\x44\x61tabase\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\"\x16\n\x14ListDatabasesRequest\"\xc9\x01\n\x15ListDatabasesResponse\x12\x41\n\x04meta\x18\x01 \x03(\x0b\x32-.migrieren.v1.ListDatabasesResponse.MetaEntryR\x04meta\x12\x34\n\tdatabases\x18\x02 \x03(\x0b\x32\x16.migrieren.v1.DatabaseR\tdatabases\x1a\x37\n\tMetaEntry\x12\x10\n\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n\x05value\x18\x02 \x01(\tR\x05value:\x02\x38\x01*\x86\x01\n\x0eMigrationState\x12\x1f\n\x1bMIGRATION_STATE_UNSPECIFIED\x10\x00\x12\x1d\n\x19MIGRATION_STATE_UNAPPLIED\x10\x01\x12\x19\n\x15MIGRATION_STATE_CLEAN\x10\x02\x12\x19\n\x15MIGRATION_STATE_DIRTY\x10\x03*s\n\x12MigrationDirection\x12#\n\x1fMIGRATION_DIRECTION_UNSPECIFIED\x10\x00\x12\x1c\n\x18MIGRATION_DIRECTION_NONE\x10\x01\x12\x1a\n\x16MIGRATION_DIRECTION_UP\x10\x02\x32\xb7\x03\n\x07Service\x12H\n\x07Migrate\x12\x1c.migrieren.v1.MigrateRequest\x1a\x1d.migrieren.v1.MigrateResponse\"\x00\x12`\n\x0f\x41pplyMigrations\x12$.migrieren.v1.ApplyMigrationsRequest\x1a%.migrieren.v1.ApplyMigrationsResponse\"\x00\x12]\n\x0ePlanMigrations\x12#.migrieren.v1.PlanMigrationsRequest\x1a$.migrieren.v1.PlanMigrationsResponse\"\x00\x12\x45\n\x06Status\x12\x1b.migrieren.v1.StatusRequest\x1a\x1c.migrieren.v1.StatusResponse\"\x00\x12Z\n\rListDatabases\x12\".migrieren.v1.ListDatabasesRequest\x1a#.migrieren.v1.ListDatabasesResponse\"\x00\x42\x45Z3github.com/alexfalkowski/migrieren/api/migrieren/v1\xea\x02\rMigrieren::V1b\x06proto3"
 
 pool = ::Google::Protobuf::DescriptorPool.generated_pool
 pool.add_serialized_file(descriptor_data)
@@ -18,11 +18,15 @@ module Migrieren
     ApplyMigrationsRequest = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.ApplyMigrationsRequest").msgclass
     ApplyMigrationsResponse = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.ApplyMigrationsResponse").msgclass
     MigrationStatus = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.MigrationStatus").msgclass
+    MigrationPlan = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.MigrationPlan").msgclass
+    PlanMigrationsRequest = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.PlanMigrationsRequest").msgclass
+    PlanMigrationsResponse = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.PlanMigrationsResponse").msgclass
     StatusRequest = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.StatusRequest").msgclass
     StatusResponse = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.StatusResponse").msgclass
     Database = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.Database").msgclass
     ListDatabasesRequest = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.ListDatabasesRequest").msgclass
     ListDatabasesResponse = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.ListDatabasesResponse").msgclass
     MigrationState = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.MigrationState").enummodule
+    MigrationDirection = ::Google::Protobuf::DescriptorPool.generated_pool.lookup("migrieren.v1.MigrationDirection").enummodule
   end
 end
